+++ lib/map/empty_map.dart
@@ -1,6 +1,8 @@
-import 'package:bonfire/map/world_map.dart';
+import 'package:bonfire/bonfire.dart';
 
 class EmptyWorldMap extends WorldMap {
-  EmptyWorldMap({double tileSizeToUpdate = 0})
-      : super([], tileSizeToUpdate: tileSizeToUpdate);
+  EmptyWorldMap({double tileSizeToUpdate = 0, Vector2? size})
+      : super([
+          if (size != null) TileModel(x: size.x, y: size.y, width: 1, height: 1)
+        ], tileSizeToUpdate: tileSizeToUpdate);
 }